Manual overlay PoC
This doesn't affect the current zaza overlay support. It would be
a different way to use overlays.

Rename bionic-train.yaml to bionic-base.yaml and add bionic-train.yaml

The idea is to be able to specify overlays like this for
functest-run-suite (tests/tests.yaml):

smoke_bundles:
  - bionic-base # base bundle
    - bionic-queens # overlay bundle 1
    - bionic-rocky  # overlay bundle 2
    - bionic-train  # overlay bundle 3

and like this in with functest-deploy:

functest-deploy -b bionic-base -o bionic-train